Moved to existing shared folder instead of components
