+++ src/analysis/retail/monk/brewmaster/modules/problems/PurifyingBrew/analyzer.ts
@@ -1,6 +1,6 @@
 import SPELLS from 'common/SPELLS';
 import talents from 'common/TALENTS/monk';
-import type { Problem } from 'interface/guide/components/ProblemList';
+import type { Problem } from 'interface/guide/shared/ProblemList';
 import Analyzer, { Options } from 'parser/core/Analyzer';
 import EventFilter, { SELECTED_PLAYER } from 'parser/core/EventFilter';
 import Events, {
